+++ src/use/router/index.ts
@@ -10,6 +10,7 @@ import PageNotFound from "@/views/errors/404.vue";
 import { createRouter, createWebHistory } from "vue-router";
 const LoginView = () => import("@/views/common/Login.vue");
 const StudentCourseView = () => import("@/views/student/StudentCourseList.vue");
+const StudentCourseInfoPage = () => import("@/views/student/CourseInfoPage.vue");
 const AllCourseView = () => import("@/views/common/CourseList.vue");
 const AdminAccountListView = () => import("@/views/admin/AdminAccountList.vue");
 const CourseForm = () => import("@/views/shared/EditCreateCourseForm.vue");
@@ -21,6 +22,7 @@ const Settings = () => import("@/views/common/Settings.vue");
 const StudentImmatricultaion = () => import("@/components/study/Immatriculation.vue");
 const OperationsArchive = () => import("@/views/common/OperationsArchive.vue");
 const AllOperationsPage = () => import("@/views/admin/AllOperations.vue");
+const DPA = () => import("@/views/common/DPA.vue");
 
 const routerHistory = createWebHistory(process.env.BASE_URL);
 const suffix: string = " | UC4";
@@ -82,6 +84,36 @@ const router = createRouter({
             },
         },
 
+        {
+            path: "/courses/my-courses",
+            name: "student.my-courses",
+            component: StudentCourseView,
+            props: { isMyCoursesPage: true },
+            meta: {
+                title: "My Courses" + suffix,
+                roles: ["Student"],
+            },
+        },
+        {
+            path: "/courses/:courseId/drop",
+            name: "student.course.drop",
+            component: StudentCourseInfoPage,
+            props: { isRegistered: true },
+            meta: {
+                title: "Course" + suffix,
+                roles: ["Student"],
+            },
+        },
+        {
+            path: "/courses/:courseId/join",
+            name: "student.course.join",
+            component: StudentCourseInfoPage,
+            props: { isRegistered: false },
+            meta: {
+                title: "Course" + suffix,
+                roles: ["Student"],
+            },
+        },
         {
             path: "/course-management",
             name: "lecturer.myCourses",
@@ -229,6 +261,14 @@ const router = createRouter({
             meta: {
                 title: "All Operations" + suffix,
                 roles: ["Admin"],
+            },
+        },
+        {
+            path: "/data-protection-agreement",
+            name: "dpa",
+            component: DPA,
+            meta: {
+                title: "Data Protection Agreement" + suffix,
             },
         },
         {