+++ src/use/store/mutations.ts
@@ -5,6 +5,7 @@ import Student from "@/api/api_models/user_management/Student";
 import { MutationTree } from "vuex";
 import { MutationTypes } from "./mutation-types";
 import defaultState, { State } from "./state";
+import Configuration from "@/api/api_models/configuration_management/Configuration";
 
 export type Mutations<S = State> = {
     [MutationTypes.SET_USER](state: S, payload: Student | Lecturer | Admin): void;
@@ -18,6 +19,8 @@ export type Mutations<S = State> = {
     [MutationTypes.FORCE_UPDATE_PROFILE_PICTURE](state: S): void;
     [MutationTypes.SET_HAS_CERTIFICATE](state: S, payload: boolean): void;
     [MutationTypes.RESET_STATE](state: S, payload: void): void;
+    [MutationTypes.SET_CONFIGURATION](state: S, payload: Configuration): void;
+    [MutationTypes.SET_VALIDATION](state: S, payload: any): void;
     [MutationTypes.ADD_OPERATION_APPROVAL](state: S, payload: string): void;
     [MutationTypes.ADD_OPERATION_REJECTION](state: S, payload: string): void;
     [MutationTypes.CLEAR_TREATED_OPERATIONS](state: S, payload: void): void;
@@ -56,6 +59,12 @@ export const mutations: MutationTree<State> & Mutations = {
     },
     [MutationTypes.RESET_STATE](state: State) {
         Object.assign(state, defaultState);
+    },
+    [MutationTypes.SET_CONFIGURATION](state: State, payload: Configuration) {
+        state.configuration = payload;
+    },
+    [MutationTypes.SET_VALIDATION](state: State, payload: any) {
+        state.validation = payload;
     },
     [MutationTypes.ADD_OPERATION_APPROVAL](state: State, payload: string) {
         state.treatedOperations.approved.push(payload);