+++ src/use/store/state.ts
@@ -2,6 +2,7 @@ import Certificate from "@/api/api_models/certificate_management/Certificate";
 import Admin from "@/api/api_models/user_management/Admin";
 import Lecturer from "@/api/api_models/user_management/Lecturer";
 import Student from "@/api/api_models/user_management/Student";
+import Configuration from "@/api/api_models/configuration_management/Configuration";
 
 export const state = {
     apiUrl: "",
@@ -13,6 +14,8 @@ export const state = {
     encryptPrivateKeyModal: {} as any,
     certificate: {} as Certificate,
     hasCertificate: false,
+    validation: {},
+    configuration: {} as Configuration,
     treatedOperations: {
         approved: [] as string[],
         rejected: [] as string[],