+++ tests/e2e/navbarMobile.spec.ts
@@ -153,6 +153,10 @@ describe("student", () => {
         //all courses
         cy.get("#nav_mobile_student_menu_courses_all").click();
         cy.url().should("contain", "courses");
+
+        //admitted courses
+        cy.get("#nav_mobile_student_menu_courses_admitted").click();
+        cy.url().should("contain", "/my-courses");
     });
 
     it("study", () => {