+++ tests/fixtures/CourseType.ts
@@ -2,5 +2,4 @@ export enum CourseType {
     LECTURE = "Lecture",
     SEMINAR = "Seminar",
     PG = "Project Group",
-    NONE = "",
 }